+++ CHANGELOG.md
@@ -329,5 +329,6 @@
 ## 0.9.8
 
 - Hot-Fix wheel install (See [issue 196](https://github.com/GRAAL-Research/deepparse/issues/196)).
+- Starting now, we also include model weights release in the GitHub release.
 
 ## dev