+++ CHANGELOG.md
@@ -334,3 +334,4 @@
 ## dev
 
 - Add version to Seq2Seq and AddressParser.
+-